% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_github_info.R
\name{backup_github_info}
\alias{backup_github_info}
\title{Copy all Issues and PRs for a user or organization}
\usage{
backup_github_info(name, save_location, entity_type = c("user", "org"))
}
\arguments{
\item{name}{Name of github user or organization (e.g., "cbedwards-dfw", "FRAMverse"). Character atomic.}

\item{save_location}{Filepath for directory to save issue and pulls to.}

\item{entity_type}{Either "user" or "org" to identify if \code{name} is describing a github user or github organization. Defaults to "user"}
}
\value{
Invisibly returns a tibble with:
\itemize{
\item \verb{$repo_name} = name of repository
\item \verb{$full_name} = full github name (\{user or org\}/\{repo name\})
\item \verb{$issues_path} = filepath to the json of issues
\item \verb{$pulls_path} = filepath to the json of pulls
\item \verb{$issues_status}, \verb{$pulls_status} = Did the backup work? "success" or "error"
\item \verb{$issues_error_msg}, \verb{$pulls_error_msg} = error messages on attempting to back up (NA if there was no error.)
}
}
\description{
To provide a backup on github failure, this function streamlines downloading all github Issues and Pull Requests of all repositories of a focal user or organization into local jsons.
}
\examples{
\dontrun{
## backing up all issues and PRs for repos hosted by `cbedwards-dfw`
out <- backup_github_info("cbedwards-dfw", "user", save_location = "C:/Repos/cbe_github_backups")
}
}
