% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_package_manuals.R
\name{build_package_manual}
\alias{build_package_manual}
\title{Build manual (and optionally vignettes) for individual repository}
\usage{
build_package_manual(
  repo_address,
  output_path,
  build_vignettes = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{repo_address}{Github repository "full name" of form \code{{user}/{repository}}, as in \code{"FRAMverse/framrsquared"}.
Also accepts full url (e.g., \verb{"https://github.com/FRAMverse/framrsquared/}) Character atomic.}

\item{output_path}{Filepath for a directory into which to save manuals. Character atomic.}

\item{build_vignettes}{Attempt to build vignettes for any package that has them? Logical,
defaults to \code{FALSE}.}

\item{verbose}{Print progress statements? Logical, defaults to \code{TRUE}.}
}
\value{
Nothing
}
\description{
For an R package stored on Github, compile a pdf reference manual and optionally
build any vignettes. Requires LaTeX to run. Under the hood, downloads temporary copy of github repo,
compiles document(s), then deletes repository.
}
\examples{
\dontrun{
build_package_manual(repo = "FRAMverse/framrsquared",
  output_path = "C:/Repos/test_manuals",
  build_vignettes = TRUE)
}
}
