% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package_network.R
\name{build_package_network}
\alias{build_package_network}
\title{Build edges of all function dependencies in package}
\usage{
build_package_network(package_name, include = "package only")
}
\arguments{
\item{package_name}{package name as character string}

\item{include}{Should only package functions be returned ("package only"),
or all functions except for base R operators ("no operators"),
or all functions ("all")?}
}
\value{
dataframe of dependency edges, with \verb{$from} and \verb{$to}
}
\description{
Build edges of all function dependencies in package
}
\examples{
build_package_network("codeInspectr")
}
