% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_package_manuals.R
\name{build_vignettes}
\alias{build_vignettes}
\title{Create vignettes for R package}
\usage{
build_vignettes(full_name, repo_dir, output_path, verbose)
}
\arguments{
\item{full_name}{Github repository "full name" of form \code{{user}/{repository}}, as in \code{"FRAMverse/framrsquared"}. Character atomic.}

\item{repo_dir}{Temporary filepath for cloned github repository}

\item{output_path}{Filepath for a directory into which to save manuals. Character atomic.}

\item{verbose}{Print progress statements? Logical, defaults to \code{TRUE}.}
}
\value{
Character vector of vignette filepaths (if successful and vignettes exist) or empty character string
}
\description{
If the git repository for an R package contains vignettes, builds them and saves them in \code{output_path},
with prefix matching the package name.
}
