% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_package_manuals.R
\name{compile_vignettes}
\alias{compile_vignettes}
\title{Create vignettes for R package}
\usage{
compile_vignettes(repo_address, repo_dir, output_path, verbose)
}
\arguments{
\item{repo_address}{Github repository "full name" of form \code{{user}/{repository}}, as in \code{"FRAMverse/framrsquared"}.
Also accepts full url (e.g., \verb{"https://github.com/FRAMverse/framrsquared/}) Character atomic.}

\item{repo_dir}{Temporary filepath for cloned github repository}

\item{output_path}{Filepath for a directory into which to save manuals. Character atomic.}

\item{verbose}{Print progress statements? Logical, defaults to \code{TRUE}.}
}
\value{
Character vector of vignette filepaths (if successful and vignettes exist) or empty character string
}
\description{
If the git repository for an R package contains vignettes, builds them and saves them in \code{output_path},
with prefix matching the package name.
}
