% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{find_function_dependencies}
\alias{find_function_dependencies}
\title{List the functions in a package that depend on a function}
\usage{
find_function_dependencies(package_name, fun, include = "package only")
}
\arguments{
\item{package_name}{package name as character string}

\item{fun}{function name as character string}

\item{include}{Should only package functions be returned ("package only"),
or all functions except for base R operators ("no operators"),
or all functions ("all")?}
}
\value{
character vector of function names
}
\description{
Returns a character vector of all functions in package \code{package_name} that depend on function \code{fun}.
}
\examples{
find_function_dependencies("codeInspectr", "find_recursive_dependencies")
body(codeInspectr::find_recursive_dependencies)
}
