% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{find_recursive_dependencies}
\alias{find_recursive_dependencies}
\title{Identify all functions that are called by the specified function, and all functions called by \emph{them}, etc}
\usage{
find_recursive_dependencies(
  package_name,
  fun,
  pattern_ignore = NULL,
  pattern_dont_recurse = NULL
)
}
\arguments{
\item{package_name}{Name of pattern. Character vector, length 1}

\item{fun}{Name of function. Character vector, length 1}

\item{pattern_ignore}{NULL or character vector of regex patterns. If present, recursive_deps will ignore any functions that match the pattern(s) character vector. Can take multiple patterns.}

\item{pattern_dont_recurse}{NULL or character vector of regex patterns. If present, recursive_deps will list functions that match the pattern(s), but will not recurse into those functions.. Can take multiple patterns.}
}
\value{
character vector
}
\description{
Recursively walk downwards through function calls. Two options to reduce traversal of functions that may not be of interest (e.g., "validate_.*" functions): \code{pattern_ignore} will completely ignore functions matching a
}
\examples{
find_recursive_dependencies("purrr", "map")
}
