% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package_network.R
\name{find_reverse_dependencies}
\alias{find_reverse_dependencies}
\title{List functions that call the specified functions}
\usage{
find_reverse_dependencies(package_name, fun, include = "package only")
}
\arguments{
\item{package_name}{package name as character string}

\item{fun}{function name as character string}

\item{include}{Should only package functions be returned ("package only"),
or all functions except for base R operators ("no operators"),
or all functions ("all")?}
}
\value{
character vector
}
\description{
To search for functions that are dependent on a non-package function (e.g.,
tracking down potential issues if a function from another package changes),
set \code{include} to "all".
}
\examples{
find_reverse_dependencies("codeInspectr", "package_functions")
}
