% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reporter.R
\name{get_branch_activity}
\alias{get_branch_activity}
\title{Return active branches and the most recent updates of github repository}
\usage{
get_branch_activity(repo_address)
}
\arguments{
\item{repo_address}{Github repository address, of the form "user/repository",
as in \code{"FRAMverse/framrsquared"}. Does not need to be an R package.}
}
\value{
Dataframe with \verb{$branch} (branch name) and \verb{$most_recent_update} (date of most recent commit)
}
\description{
Summarizes recent activity on a remote github repository, listing active branches and their
most recent updates. #' Will only process the first 100 branches of a repository.
Repository does not need to be an R package, or even contain R code. Will not
list activity for "gh-pages", which is the branch used for pkgdown.
}
\examples{
get_branch_activity("FRAMverse/framrsquared")
get_branch_activity("tidyverse/dplyr")
}
