% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reporter.R
\name{get_issues}
\alias{get_issues}
\title{Return summary of open Github Issues}
\usage{
get_issues(repo_address, max_char = 97)
}
\arguments{
\item{repo_address}{Github repository address, of the form "user/repository",
as in \code{"FRAMverse/framrsquared"}. Does not need to be an R package.}

\item{max_char}{Maximum number of characters to return in the "body" column.}
}
\value{
dataframe with \verb{$title}, \verb{$body}, and \verb{$date} of open issues.
}
\description{
Return summary of open Github Issues
}
\examples{
get_issues("FRAMverse/framrsquared")
get_issues("tidyverse/dplyr")
}
