% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reporter.R
\name{get_pull_requests}
\alias{get_pull_requests}
\title{Return summary of open Github Pull Requests}
\usage{
get_pull_requests(repo_address, max_char = 97)
}
\arguments{
\item{repo_address}{Github repository address, of the form "user/repository",
as in \code{"FRAMverse/framrsquared"}. Does not need to be an R package.}

\item{max_char}{Maximum number of characters to return in the "body" column.}
}
\value{
dataframe with \verb{$title}, \verb{$body}, and \verb{$date} of open pull requests.
}
\description{
Including title, initial text of body, and date
}
\examples{
get_pull_requests("FRAMverse/framrsquared")
get_pull_requests("tidyverse/dplyr")
}
