% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_info.R
\name{get_raw_pulls}
\alias{get_raw_pulls}
\title{Obtain Github pull requests from a repository}
\usage{
get_raw_pulls(repo_address = NULL, entity_name = NULL, repo_name = NULL)
}
\arguments{
\item{repo_address}{Github identifier in the form of \{name\}/\{repository\}, as in "FRAMverse/codeInspectr".
Can also be in the form of full URL (e.g. "https://github.com/FRAMverse/framrsquared/").
Character string, defaults to NULL. If not provided, users must instead provide the same information in \code{entity_name} and \code{repo_name}.}

\item{entity_name}{Only needed if \code{repo_address} is not provided. Name of a github user or organization repository, e.g., "FRAMverse". Character atomic, defaults to \code{NULL}.}

\item{repo_name}{Only needed if \code{repo_address} is not provided. Name of individual repository, e.g., "codeInspectr". Character atomic, defaults to \code{NULL}.}
}
\value{
A list of class \code{gh_response} containing all pull requests from a repository.
}
\description{
Returns github issues as a list. Github issues themselves are structured as jsons.
}
\examples{
# temp <- get_raw_pulls("cbedwards-dfw/codeInspectr")
}
