% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_repos.R
\name{list_repos}
\alias{list_repos}
\title{List all repositories of a Github user or organization}
\usage{
list_repos(
  name,
  entity_type = c("user", "org"),
  repo_type = c("all", "public", "private", "forks")
)
}
\arguments{
\item{name}{Name of user or organization (e.g., "FRAMverse" or "cbedwards-dfw"). Character atomic.}

\item{entity_type}{Is this a Github User ("user") or organization ("org")? Character atomic, defaults to "user".}

\item{repo_type}{List "all" repos, "public" repos, "private" repos, or only "fork" repos? Character atomic, defaults to "all". Private repos will only be provided if you have access to the private repos of that user or org.}
}
\value{
Data frame of repositories with columns
\itemize{
\item \verb{$name}: repository name, e.g., "codeInspectr"
\item \verb{$full_name}: \{user/org\}/\{repository name\}, e.g., "FRAMverse/codeInspectr"
\item \verb{$html_url}: web address of repository
\item \verb{$description}: description listed for the repository.
\item \verb{$stars}: number of followers of the repository
\item \verb{$language}: programming language(s) of the repository.
}
}
\description{
Creates a dataframe of repositories and associated info for the
user or organization specified
}
\examples{
list_repos("FRAMverse", entity_type = "org", repo_type = "public")

}
