% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{package_functions}
\alias{package_functions}
\title{List the functions of a package}
\usage{
package_functions(package_name)
}
\arguments{
\item{package_name}{Name of package as character string}
}
\value{
character vector of function names
}
\description{
List the functions of a package
}
\examples{
package_functions("codeInspectr")
}
