% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{plot_function_dependencies}
\alias{plot_function_dependencies}
\title{Plot dependency map of a function}
\usage{
plot_function_dependencies(
  package_name,
  fun,
  pattern_ignore = NULL,
  pattern_dont_recurse = NULL
)
}
\arguments{
\item{package_name}{Name of pattern. Character vector, length 1}

\item{fun}{Name of function. Character vector, length 1}

\item{pattern_ignore}{NULL or character vector of regex patterns. If present, recursive_deps will ignore any functions that match the pattern(s) character vector. Can take multiple patterns.}

\item{pattern_dont_recurse}{NULL or character vector of regex patterns. If present, recursive_deps will list functions that match the pattern(s), but will not recurse into those functions.. Can take multiple patterns.}
}
\value{
visNetwork plot
}
\description{
Plot dependency map of a function
}
\examples{
plot_function_dependencies("purrr", "map")
plot_function_dependencies("purrr", "map_df")
}
