% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_github_info.R
\name{process_repo}
\alias{process_repo}
\title{Helper function to download issues/pulls of single repo}
\usage{
process_repo(repo_name, full_name, save_location)
}
\arguments{
\item{repo_name}{Name of repository (e.g. "codeInspectr")}

\item{full_name}{Full github name (e.g., "FRAMverse/codeInspectr")}

\item{save_location}{Filepath for directory to save issue and pulls to.}
}
\value{
Tibble with filepaths and success/failure info.
}
\description{
Intended for use in purr workflow in \code{\link[=backup_github_info]{backup_github_info()}}.
}
