% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeInspectr-package.R
\docType{package}
\name{codeInspectr-package}
\alias{codeInspectr}
\alias{codeInspectr-package}
\title{codeInspectr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/codeInspectr}
  \item \url{https://cbedwards-dfw.github.io/codeInspectr/}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/codeInspectr/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
