% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_package_manuals.R
\name{compile_package_manuals}
\alias{compile_package_manuals}
\title{Save R package documentation as PDFs}
\usage{
compile_package_manuals(
  output_path,
  name,
  entity_type = c("user", "org"),
  build_vignettes = FALSE,
  repo_type = c("all", "public", "private", "forks"),
  verbose = TRUE
)
}
\arguments{
\item{output_path}{Filepath for a directory into which to save manuals. Character atomic.}

\item{name}{Name of user or organization (e.g., "FRAMverse" or "cbedwards-dfw"). Character atomic.}

\item{entity_type}{Is this a Github User ("user") or organization ("org")? Character atomic, defaults to "user".}

\item{build_vignettes}{Attempt to build vignettes for any package that has them? Logical,
defaults to \code{FALSE}.}

\item{repo_type}{List "all" repos, "public" repos, "private" repos, or only "fork" repos? Character atomic, defaults to "all". Private repos will only be provided if you have access to the private repos of that user or org.}

\item{verbose}{Print progress statements? Logical, defaults to \code{TRUE}.}
}
\value{
Nothing
}
\description{
Creates PDF manuals for all R packages of a given user or organization. Manuals are created with
devtools::build_manual. This requires a version of LaTeX installed. If you don't have a version
installed, you can install it with \code{tinytex::install_tinytex()}. Optionally can also build
vignettes for each package. Successfully building vignettes may require additional packages be
installed (whatever packages the vignettes depend on).
}
\examples{
\dontrun{
# build manuals and vignettes for all R packages in the FRAMverse organization.
compile_package_manuals("C:/Repos/manuals",
                        name = "FRAMverse", entity_type = "org", build_vignettes = TRUE)
}
}
