% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_package.R
\name{is_r_package}
\alias{is_r_package}
\title{Test if repository is an R package}
\usage{
is_r_package(full_name, default_branch = "main")
}
\arguments{
\item{full_name}{Repository name including user or organization prefix, e.g., "FRAMverse/framrsquared". Character atomic}

\item{default_branch}{Name of the default/primary repository. Character atomic, defaults to "main". Older repositories often use "master" as the default repo name.}
}
\value{
Logical: \code{TRUE} if the repo is an R package, \code{FALSE} if it is not.
}
\description{
Uses the existence of a DESCRIPTION file and the first line being "Package:..." to decide if
the provided repository is an R package. Accesses first few lines of DESCRIPTION in \verb{raw.githubusercontent.com/}
for speed.
}
\examples{
# framrsquared is an R package
is_r_package("FRAMverse/framrsquared")
# the snippets repository is not.
is_r_package("FRAMverse/snippets")
}
