% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reporter.R
\name{summarize_repository}
\alias{summarize_repository}
\title{Summarize information about a github R package}
\usage{
summarize_repository(repo_address, max_char = 70)
}
\arguments{
\item{repo_address}{Github repository address for an R package, of the form "user/repository",
as in \code{"FRAMverse/framrsquared"}}

\item{max_char}{Maximum number of characters to return in the "body" column.}
}
\value{
List summarizing github repository information.
}
\description{
Summarize information about a github R package
}
\examples{
summarize_repository("FRAMverse/framrsquared")
summarize_repository("cbedwards-dfw/xldiff")
}
\seealso{
\code{\link[=get_issues]{get_issues()}}, \code{\link[=get_pull_requests]{get_pull_requests()}}, \code{\link[=get_branch_activity]{get_branch_activity()}}
}
